Relocate $literal project disambiguation tests to integration level
Move tests from misc/literal/test_literal_project_disambiguation.py:
- Project disambiguation tests (0, 1, true, false) and new -1 case move
  to core/operator/test_operator_combination_literal_stages.py
- Dollar-sign $eq comparison tests move to
  expressions/test_expressions_combination_literal.py

Delete the original file since it tested cross-feature interactions that
belong at the parent level per folder structure guidelines.
